merge Dashboard dpa

+++ app/Http/Controllers/DpaController.php
@@ -4,44 +4,39 @@ namespace App\Http\Controllers;
 
 use App\Models\Dpa;
 use App\Models\User;
+use App\Models\BimbinganStudy;
 use Illuminate\Http\Request;
-use App\Models\StudentType;
 use App\Models\StudyProgram;
 use Illuminate\View\View;
 use Illuminate\Http\RedirectResponse;
 use App\Http\Requests\dpa\DpaCreateRequest;
 use App\Http\Requests\dpa\DpaUpdateRequest;
+use App\Models\Student;
 use App\Models\Ukt;
+use Illuminate\Support\Facades\Auth;
 
 class DpaController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index():View
     {
         return view('dpa.data')->with([
-            'dpa' => Dpa::paginate(20),
+            'dpa' => Dpa::latest()->paginate(30),
         ]);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create():View
     {
-        return view('dpa.create');
+        $study_program = StudyProgram::all();
+        return view('dpa.create',  compact('study_program'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(DpaCreateRequest $request):RedirectResponse
     {
-        
+
         // Simpan akun user DPA
         $name = $request->name;
         $email = $request->email;
+        $study_program = $request->study_program_id;
         $password = bcrypt('admin');
         $role = 'DPA';
         $this->addUser($name, $email, $password, $role);
@@ -51,6 +46,7 @@ class DpaController extends Controller
         Dpa::create([
             'name' => $name,
             'email' => $email,
+            'study_program_id' => $study_program,
             'user_id' => $user_id['id'],
         ]);
 
@@ -73,6 +69,7 @@ class DpaController extends Controller
         //
         return view('dpa.edit')->with([
             'dpa' => Dpa::findOrFail($id),
+            "study_program" => StudyProgram::all(),
         ]);
     }
 
@@ -93,7 +90,7 @@ class DpaController extends Controller
             'name' => $name,
             'email' => $email,
         ]);
-        
+
         return redirect()->route('dpa.index')->with(['success' => 'Data berhasil diupdate']);
     }
 
@@ -103,10 +100,19 @@ class DpaController extends Controller
     public function destroy($id)
     {
         $dpa = Dpa::findOrFail($id);
-        $user_id = $dpa['user_id'];
-        $user = User::findOrFail($user_id);
-        $dpa->delete();
-        $user->delete();
+        $student = Student::where('dpa_id', $id)->exists();
+
+        if (!$student) {
+            $user_id = $dpa['user_id'];
+            $user = User::findOrFail($user_id);
+            $dpa->delete();
+            $user->delete();
+        } else{
+            return redirect()->route('dpa.index')->with(['warning' => 'Data DPA masih terhubung dengan data Mahasiswa']);
+        }
+        return view('dpa.data')->with([
+            'dpa' => Dpa::latest()->paginate(30),
+        ]);
     }
 
     public function addUser($name, $email, $password, $role)
@@ -117,5 +123,159 @@ class DpaController extends Controller
             'password' => $password,
             'role' => $role,
         ]);
+    }
+
+    public function getMahasiswa(Request $request): View
+    {
+        $dpa_id = $request->dpa_id;
+        $tahunAjaran = $request->year;
+        $semester = $request->semester;
+        $lbs_id = $request->input('id');
+
+        if (empty($dpa_id)) {
+            $user_id = Auth::user()->id;
+            $dpa = Dpa::where('user_id', $user_id)->first();
+            $dpa_id = $dpa->id;
+        }
+
+        if (empty($tahunAjaran) || empty($semester)) {
+            $tahunAjaran = date('Y');
+            $semester = "GASAL";
+        }
+
+        $dpa = Dpa::where('id', $dpa_id)->first();
+        $students = Student::where('dpa_id', $dpa_id)->get();
+
+        if (!empty($lbs_id)) {
+            $this->setujuKrs($lbs_id);
+        }
+
+        $data = $this->getBimbinganStudi($students, $tahunAjaran, $semester);
+
+        return view('dpa.daftarmahasiswa')->with([
+            'data' => $data,
+            'tahunAjaran' => [$tahunAjaran, $semester],
+            'dpa' => $dpa,
+
+        ]);
+    }
+
+    public function export(Request $request)
+    {
+        $dpa_id = $request->dpa_id;
+        $tahunAjaran = $request->year;
+        $semester = $request->semester;
+
+        if (empty($tahunAjaran) || empty($semester)) {
+            $tahunAjaran = date('Y');
+            $semester = "GASAL";
+        }
+
+        if (empty($dpa_id)) {
+            $user_id = Auth::user()->id;
+            $dpa = Dpa::where('user_id', $user_id)->first();
+            $dpa_id = $dpa->id;
+        }
+
+        $dpa = Dpa::where('id', $dpa_id)->first();
+        $students = Student::where('dpa_id', $dpa_id)->get();
+
+        $data = $this->getBimbinganStudi($students, $tahunAjaran, $semester);
+
+        return view('report.printformat.daftarmahasiswa')->with([
+            'data' => $data,
+            'tahunAjaran' => [$tahunAjaran, $semester],
+            'dpa' => $dpa,
+            'title' => "Lembar Bimbingan Studi",
+            'today' => date('d F Y', strtotime(date('Y-m-d'))),
+        ]);
+    }
+
+    public function setujuKrs($id)
+    {
+        $uktController = new UktController;
+
+        $bimbinganStudy = BimbinganStudy::where('id', $id)->first();
+        $bimbinganStudy->status = "Aktif";
+        $bimbinganStudy->save();
+
+        $payment = Ukt::where('lbs_id', $id)->first();
+        if ($payment->keterangan == "UTS") {
+            $payment->exam_uts_id = $uktController->createExamCard($payment->students_id, "UTS", $payment->semester, $payment->year);
+        } elseif ($payment->keterangan == "UAS") {
+            $payment->exam_uts_id = $uktController->createExamCard($payment->students_id, "UTS", $payment->semester, $payment->year);
+            $payment->exam_uas_id = $uktController->createExamCard($payment->students_id, "UAS", $payment->semester, $payment->year);
+        }
+        $payment->save();
+    }
+
+    public function getBimbinganStudi($students, $tahunAjaran, $semester)
+    {
+        $data = [];
+        foreach ($students as $item) {
+            $wisuda = Ukt::where('students_id', $item->id)->where('type', "WISUDA")->exists();
+
+            if ($semester == "GASAL") {
+                $semesterStudent = (($tahunAjaran - $item->force) * 2) + 1;
+            } elseif ($semester == "GENAP") {
+                $semesterStudent = (($tahunAjaran - $item->force) * 2);
+            }
+
+            if ($semesterStudent >= 1 && !$wisuda) {
+                $bimbinganStudy = BimbinganStudy::where('students_id', $item->id)->where('year', $tahunAjaran)->where('semester', $semester)->first();
+                if (empty($bimbinganStudy)) {
+                    $status = "Tidak Aktif";
+                    $lbs_id = null;
+                } else {
+                    $lbs_id = $bimbinganStudy->id;
+                    $status = $bimbinganStudy->status;
+                }
+
+                $data[$item->id] = [
+                    'id' => $item->id,
+                    'nim' => $item->nim,
+                    'name' => $item->name,
+                    'semester' => $semesterStudent,
+                    'lbs_id' => $lbs_id,
+                    'status' => $status
+                ];
+            }
+        }
+
+        return $data;
+    }
+
+    public function getDetailDpa(Request $request, string $id):RedirectResponse
+    {
+        $dpa = Dpa::findOrFail($id);
+        $dpa->update($request->all());
+
+        $user_id = $dpa['user_id'];
+        $user = User::findOrFail($user_id);
+        $user->update([
+            'status' => 'Aktif',
+        ]);
+
+        return redirect()->route('dpa.index')->with(['success' => 'Data berhasil diupdate']);
+    }
+
+    public function getData($datepicker, $filter)
+    {
+        $date = date('Y-m');
+        $formattedDate = date('F Y');
+
+        if (!empty($datepicker)) {
+            if ($filter == 'month') {
+                $parsedDate = \DateTime::createFromFormat('m-Y', $datepicker);
+                $date = $parsedDate->format('Y-m');
+                $formattedDate = $parsedDate->format('F Y');
+            } elseif ($filter == 'year') {
+                $parsedDate = \DateTime::createFromFormat('Y', $datepicker);
+                $date = $parsedDate->format('Y');
+                $formattedDate = $parsedDate->format('Y');
+            }
+        }
+
+        return [$date, $formattedDate];
     }
 }
